refactoring stuff around
