+++ clients/apps/web/src/components/Benefit/BenefitSections.tsx
